Move resolution module to `e3sm/init` framework
