+++ polaris/tasks/e3sm/init/topo/__init__.py
@@ -1,15 +1,1 @@
-from polaris.tasks.e3sm.init.topo.resolutions import (
-    CUBED_SPHERE_RESOLUTIONS as CUBED_SPHERE_RESOLUTIONS,
-)
-from polaris.tasks.e3sm.init.topo.resolutions import (
-    LAT_LON_RESOLUTIONS as LAT_LON_RESOLUTIONS,
-)
-from polaris.tasks.e3sm.init.topo.resolutions import (
-    format_lat_lon_resolution_name as format_lat_lon_resolution_name,
-)
-from polaris.tasks.e3sm.init.topo.resolutions import (
-    get_cubed_sphere_resolution as get_cubed_sphere_resolution,
-)
-from polaris.tasks.e3sm.init.topo.resolutions import (
-    uses_low_res_cubed_sphere as uses_low_res_cubed_sphere,
-)
+"""Topography tasks for the E3SM init task family."""
