+++ polaris/tasks/e3sm/init/topo/combine/step.py
@@ -9,8 +9,8 @@ import xarray as xr
 from mpas_tools.logging import check_call
 from pyremap import ProjectionGridDescriptor, get_lat_lon_descriptor
 
+from polaris.e3sm.init.topo import format_lat_lon_resolution_name
 from polaris.step import Step
-from polaris.tasks.e3sm.init.topo import format_lat_lon_resolution_name
 
 
 class CombineStep(Step):
